# More optimization of rev-list command.