The code was previously parsing and re-parsing the index file for each
object it was trying to retrieve from a pack, thereby defeating the
purpose of indexing it in the first place. This finally makes some
significant progress on rev-list benchmark by caching the index after
being parsed.

Before:
BenchmarkRevListHead-4                         1        4478326709 ns/op
BenchmarkRevListHeadExclude-4                  1        4360914540 ns/op
BenchmarkRevListHeadExcludeObjects-4           1        15265898270 ns/op

After:
BenchmarkRevListHead-4                      2000            797315 ns/op
BenchmarkRevListHeadExclude-4               2000            917981 ns/op
BenchmarkRevListHeadExcludeObjects-4           1        2602027452 ns/op